Rydder i app-strukturen
