+++ src/komponenter/header/header-regular/HeaderMenylinje.test.tsx
@@ -6,13 +6,13 @@ import { reducers } from 'store/reducers';
 import { mount } from 'enzyme';
 import NavLogoLenke from 'komponenter/common/nav-logo/NavLogoLenke';
 import { HeaderMenylinje } from 'komponenter/header/header-regular/HeaderMenylinje';
-import { HovedmenyKnapp } from 'komponenter/header/header-regular/common/knapper/hovedmeny-knapp/HovedmenyKnapp';
+import { HovedmenyKnapp } from 'komponenter/header/header-regular/common/meny-knapp/hovedmeny-knapp/HovedmenyKnapp';
 import { ActionType } from 'store/actions';
 import { Status } from 'api/api';
-import LoggInnKnapp from 'komponenter/header/header-regular/common/knapper/logg-inn-knapp/LoggInnKnapp';
+import LoggInnKnapp from 'komponenter/header/header-regular/common/logg-inn/LoggInnKnapp';
 import { settArbeidsflate } from 'store/reducers/arbeidsflate-duck';
 import { MenuValue } from 'utils/meny-storage-utils';
-import { VarslerKnapp } from 'komponenter/header/header-regular/common/knapper/varsler-knapp/VarslerKnapp';
+import { VarslerKnapp } from 'komponenter/header/header-regular/common/varsler/varsler-knapp/VarslerKnapp';
 import MinsideMeny from 'komponenter/header/header-regular/desktop/minside-meny/MinsideMeny';
 import { innloggetAction } from 'utils/jest/testObjects';
 import { uInnloggetAction } from 'utils/jest/testObjects';
