--- src/komponenter/header/header-regular/common/knapper/ikoner/minside-ikon/MinsideIkon.less
@@ -1,25 +1,0 @@
-@import (reference) '../../../../../../../styling-variabler.less';
-
-.minside-meny-ikon {
-    @width: 32px;
-    @height: 32px;
-    @stroke-width: 1px;
-    @color: @navBla;
-    @hode-diameter: 12px;
-    @kropp-radius: @hode-diameter;
-    @kropp-slimness: 4px;
-
-    position: relative;
-    width: @width;
-    height: @height;
-    padding: 0;
-    top: 2px;
-    margin-right: 0.25em;
-
-    .hamburger-ikon {
-        right: @kropp-radius;
-        top: 2px;
-        transform-origin: center;
-        transform: scale(0.75);
-    }
-}
