+++ src/komponenter/header/header-regular/common/logg-inn/LoggInnKnapp.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../../../styling-variabler.less';
+@import (reference) '../../../../../styling-variabler.less';
 
 // Safari fix
 .set-text-color(@mainColor, @activeColor, @hoverColor, @focusColor) {
