+++ src/komponenter/header/header-regular/common/meny-knapp/hamburger-ikon/HamburgerIkon.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../../../../styling-variabler.less';
+@import (reference) '../../../../../../styling-variabler.less';
 
 .hamburger-ikon {
     @width: 28px;
