+++ src/komponenter/header/header-regular/common/meny-knapp/hovedmeny-knapp/HovedmenyKnapp.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import MenylinjeKnapp from 'komponenter/header/header-regular/common/knapper/MenylinjeKnapp';
-import HamburgerIkon from 'komponenter/header/header-regular/common/knapper/ikoner/hamburger-ikon/HamburgerIkon';
+import MenylinjeKnapp from 'komponenter/header/header-regular/common/meny-knapp/MenylinjeKnapp';
+import HamburgerIkon from 'komponenter/header/header-regular/common/meny-knapp/hamburger-ikon/HamburgerIkon';
 import Undertittel from 'nav-frontend-typografi/lib/undertittel';
 import Tekst from 'tekster/finn-tekst';
 import { gaEvent } from 'utils/google-analytics';
