+++ src/komponenter/header/header-regular/common/varsler/varsler-knapp/varsler-ikon/VarslerIkon.less
