+++ src/komponenter/header/header-regular/common/varsler/varsler-knapp/varsler-ikon/VarslerIkon.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import BEMHelper from 'utils/bem';
 import { EtikettLiten } from 'nav-frontend-typografi';
-import './VarselIkon.less';
+import './VarslerIkon.less';
 
 type Props = {
     isOpen: boolean;
     antallUleste?: number;
 };
 
-export const VarselIkon = ({ isOpen, antallUleste = 0 }: Props) => {
+export const VarslerIkon = ({ isOpen, antallUleste = 0 }: Props) => {
     const cls = BEMHelper('varselbjelle-ikon');
 
     return (
