+++ src/komponenter/header/header-regular/desktop/minside-meny/MinsideMeny.test.tsx
@@ -12,9 +12,9 @@ import { Status } from 'api/api';
 import mockMenu from 'server/mock/menu.json';
 import { languageDuck } from 'store/reducers/language-duck';
 import { Language } from 'store/reducers/language-duck';
-import { MinsidePersonKnapp } from 'komponenter/header/header-regular/common/knapper/minside-knapper/MinsidePersonKnapp';
-import MinsideArbgiverKnapp from 'komponenter/header/header-regular/common/knapper/minside-knapper/MinsideArbgiverKnapp';
-import { MinsideKnapp } from 'komponenter/header/header-regular/common/knapper/minside-knapper/MinsideKnapp';
+import { MinsidePersonKnapp } from 'komponenter/header/header-regular/desktop/minside-meny/minside-knapper/MinsidePersonKnapp';
+import MinsideArbgiverKnapp from 'komponenter/header/header-regular/desktop/minside-meny/minside-knapper/MinsideArbgiverKnapp';
+import { MinsideKnapp } from 'komponenter/header/header-regular/desktop/minside-meny/minside-knapper/MinsideKnapp';
 import { innloggetAction } from 'utils/jest/testObjects';
 import { kbNavDummy } from 'utils/jest/testObjects';
