+++ src/komponenter/header/header-regular/desktop/minside-meny/minside-knapper/MinsideKnapp.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { minsideDropdownClassname } from 'komponenter/header/header-regular/desktop/minside-meny/MinsideMeny';
+import { minsideDropdownClassname } from '../MinsideMeny';
 import { MenuValue } from 'utils/meny-storage-utils';
 import MinsideArbgiverKnapp from './MinsideArbgiverKnapp';
 import { MinsidePersonKnapp } from './MinsidePersonKnapp';
-import 'komponenter/header/header-regular/common/knapper/minside-knapper/MinsideKnapp.less';
+import './MinsideKnapp.less';
 
 export const minsideKnappId = 'desktop-minside-knapp-id';
