+++ src/komponenter/header/header-regular/desktop/minside-meny/minside-knapper/minside-ikon/MinsideIkon.less
@@ -1,11 +1,27 @@
 @import (reference) '../../../../../../../styling-variabler.less';
 
-.hode-kropp-ikon {
+.minside-meny-ikon {
+    @width: 32px;
+    @height: 32px;
     @stroke-width: 1px;
     @color: @navBla;
     @hode-diameter: 12px;
     @kropp-radius: @hode-diameter;
     @kropp-slimness: 4px;
+
+    position: relative;
+    width: @width;
+    height: @height;
+    padding: 0;
+    top: 2px;
+    margin-right: 0.25em;
+
+    .hamburger-ikon {
+        right: @kropp-radius;
+        top: 2px;
+        transform-origin: center;
+        transform: scale(0.75);
+    }
 
     &__hode {
         border-radius: 50%;
