+++ src/komponenter/header/header-regular/desktop/sok-dropdown/sok-knapp/SokMenyIkon.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../../../../styling-variabler.less';
+@import (reference) '../../../../../../styling-variabler.less';
 
 .sok-meny-ikon {
     @width: 32px;
