+++ src/komponenter/header/header-regular/desktop/sok-dropdown/sok-knapp/SokMenyIkon.tsx
