+++ src/komponenter/header/header-simple/HeaderSimple.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import BEMHelper from 'utils/bem';
 import NavLogoLenke from 'komponenter/common/nav-logo/NavLogoLenke';
 import Navn from './navn/Navn';
-import LoggInnKnapp from 'komponenter/header/header-regular/common/knapper/logg-inn-knapp/LoggInnKnapp';
+import LoggInnKnapp from 'komponenter/header/header-regular/common/logg-inn/LoggInnKnapp';
 import { Sticky } from 'komponenter/header/header-regular/common/sticky/Sticky';
 import { GACategory } from 'utils/google-analytics';
 import './HeaderSimple.less';
