+++ src/utils/keyboard-navigation/kb-navigation-setup.ts
@@ -12,10 +12,10 @@ import { Status } from 'api/api';
 import { kbMasterNode } from 'utils/keyboard-navigation/useKbNavMain';
 import { desktopSokInputId } from 'komponenter/header/header-regular/desktop/sok-dropdown/SokDropdown';
 import { headerLogoId } from 'komponenter/header/header-regular/HeaderMenylinje';
-import { loginKnappId } from 'komponenter/header/header-regular/common/knapper/logg-inn-knapp/LoggInnKnapp';
-import { desktopSokKnappId } from 'komponenter/header/header-regular/common/knapper/sok-knapp/SokKnapp';
-import { varslerKnappId } from 'komponenter/header/header-regular/common/knapper/varsler-knapp/VarslerKnapp';
-import { minsideKnappId } from 'komponenter/header/header-regular/common/knapper/minside-knapper/MinsideKnapp';
+import { loginKnappId } from 'komponenter/header/header-regular/common/logg-inn/LoggInnKnapp';
+import { desktopSokKnappId } from 'komponenter/header/header-regular/desktop/sok-dropdown/sok-knapp/SokKnapp';
+import { varslerKnappId } from 'komponenter/header/header-regular/common/varsler/varsler-knapp/VarslerKnapp';
+import { minsideKnappId } from 'komponenter/header/header-regular/desktop/minside-meny/minside-knapper/MinsideKnapp';
 import { desktopHovedmenyKnappId } from 'komponenter/header/header-regular/desktop/hovedmeny/HovedmenyDesktop';
 
 export type KbNavConfig = {
